+++ tests/testthat/test-acceptance_ratio.R
@@ -7,32 +7,32 @@ test_that("acceptance rates are computed", {
     compute_options = set_compute_options(burnin = 10, nmc = 100),
     cl = cl
   )
-  expect_equal(mod$acceptance_ratios$alpha_acceptance, list(0.911111111111111, 0.833333333333333))
-  expect_equal(mod$acceptance_ratios$rho_acceptance, list(0.977777777777778, 0.9))
+  expect_equal(get_acceptance_ratios(mod)$alpha_acceptance, list(0.911111111111111, 0.833333333333333))
+  expect_equal(get_acceptance_ratios(mod)$rho_acceptance, list(0.977777777777778, 0.9))
 
   mod <- compute_mallows(
     data = setup_rank_data(potato_visual),
     compute_options = set_compute_options(burnin = 10, nmc = 100, alpha_prop_sd = 100),
     cl = cl
   )
-  expect_equal(mod$acceptance_ratios$alpha_acceptance, list(0.0444444444444444, 0.0333333333333333))
-  expect_equal(mod$acceptance_ratios$rho_acceptance, list(0.977777777777778, 0.988888888888889))
+  expect_equal(get_acceptance_ratios(mod)$alpha_acceptance, list(0.0444444444444444, 0.0333333333333333))
+  expect_equal(get_acceptance_ratios(mod)$rho_acceptance, list(0.977777777777778, 0.988888888888889))
 
   mod <- compute_mallows(
     data = setup_rank_data(potato_visual),
     compute_options = set_compute_options(burnin = 10, nmc = 100, alpha_prop_sd = 0.00001),
     cl = cl
   )
-  expect_equal(mod$acceptance_ratios$alpha_acceptance, list(0.988888888888889, 0.988888888888889))
-  expect_equal(mod$acceptance_ratios$rho_acceptance, list(0.922222222222222, 0.888888888888889))
+  expect_equal(get_acceptance_ratios(mod)$alpha_acceptance, list(0.988888888888889, 0.988888888888889))
+  expect_equal(get_acceptance_ratios(mod)$rho_acceptance, list(0.922222222222222, 0.888888888888889))
 
   mod <- compute_mallows(
     data = setup_rank_data(potato_visual),
     compute_options = set_compute_options(burnin = 10, nmc = 100, leap_size = 9),
     cl = cl
   )
-  expect_equal(mod$acceptance_ratios$alpha_acceptance, list(0.711111111111111, 0.777777777777778))
-  expect_equal(mod$acceptance_ratios$rho_acceptance, list(0.566666666666667, 0.511111111111111))
+  expect_equal(get_acceptance_ratios(mod)$alpha_acceptance, list(0.711111111111111, 0.777777777777778))
+  expect_equal(get_acceptance_ratios(mod)$rho_acceptance, list(0.566666666666667, 0.511111111111111))
 
   set.seed(1)
   mod1 <- compute_mallows(